move storage spec to pkg/alerting
